rest controller for generating app
