+++ src/main/java/org/jbpm/bootstrap/service/security/BootstrapWebSecurityConfig.java
@@ -30,6 +30,7 @@ public class BootstrapWebSecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
 		http
+        .headers().frameOptions().disable().and()
         .csrf().disable()
         .authorizeRequests()
             .regexMatchers("/rest/server/*").authenticated()
